Add config.js
mongo-express should come with a config.js file (exact copy of config.default.js). This way it will run out of the box for mongodb default installations, not needing to (first figure out, and then) make this copy manually. I strongly believe this friction is currently stopping a lot of people from using this tool.