Revert back to the original way of building docker
The original way of building the docker image was to use the source already in teh current directory. That is better than the way I updated it to be, which is to pull a specific version from npm. (redundant because it's already here)

This also removes the need to keep track of the latest node package version number.